-2º attempt to fix Linux Open GL ES 1.0 building

+++ source/client/gui/Screen.cpp
@@ -238,7 +238,7 @@ void Screen::renderPanoramaBlur(float f)
 {
 	glBindTexture(GL_TEXTURE_2D, m_pMinecraft->m_pTextures->loadBlankTexture("panorama_blur", 256, 256));
 #if defined(USE_GLES1_COMPATIBILITY_LAYER) && USE_GLES1_COMPATIBILITY_LAYER
-	static GLubyte buffer[256 * 256 * 4];
+	static uint32_t buffer[256 * 256 * 4];
 	glReadPixels(0, 0, 256, 256, GL_RGBA, GL_UNSIGNED_BYTE, buffer);
 	glTexSubImage2D(GL_TEXTURE_2D, 0, 0, 0, 256, 256,
 		GL_RGBA, GL_UNSIGNED_BYTE, buffer);
